+++ libosmscout/include/osmscout/Util.h
@@ -180,7 +180,7 @@ namespace osmscout {
                            size_t bufferLength,
                            char* buffer,
                            size_t& bytes);
-  extern bool DecodeNumber(const char* buffer, unsigned long& number, size_t& bytes);
+  extern bool DecodeNumber(const char* buffer, uint32_t& number, size_t& bytes);
 
   extern bool GetFileSize(const std::string& filename, long& size);
